# src/api/specs/swagger.yaml
openapi: 3.0.0
info:
  title: Next.js API
  description: API documentation for Next.js 15 application
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local server
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UnAuthorizedError:
      type: object
      properties:
        error:
          type: string
          example: Unauthorized
        message:
          type: string
          example: Please log in.
...
    NotFoundError:
      type: object
      properties:
        error:
          type: string
          example: NotFound
        message:
          type: string
          example: The target data does not exist.
...
    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: ValidationError
        errors:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
                example: email
              message:
                type: string
                example: Invalid email address
    InternalServerError:
      type: object
      properties:
        error:
          type: string
          example: InternalServerError
        message:
          type: string
          example: Server error.
    Success:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Success
    Created:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Created
...
paths:
  /api/admin/users/create-user:
    post:
      summary: Register a new user
      description: >-
        Creates a new user, hashes their password, generates an email
        verification token, and sends a verification email.
      operationId: adminCreateUser
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - firstName
                - lastName
                - country
                - email
                - password
              properties:
                firstName:
                  type: string
                  example: John
                lastName:
                  type: string
                  example: Doe
                country:
                  type: string
                  example: Bangladesh
                gender:
                  type: string
                  enum:
                    - male
                    - female
                    - other
                  example: male
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: SecureP@ssw0rd
                address:
                  type: string
                  example: 123 Main St
                role:
                  type: string
                  example: user
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Created'
        '400':
          description: Validation errors or user already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'User already exists with this email: user@example.com'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
...
  /api/admin/categories/get-category:
    post:
      summary: Fetch categories with pagination and search
      description: Retrieves categories from the database with pagination and optional search.
      operationId: adminGetCategories
      tags:
        - Admin
      security:
        - jwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                page:
                  type: integer
                  example: 1
                limit:
                  type: integer
                  example: 10
                search:
                  type: string
                  example: tech
      responses:
        '200':
          description: Successfully retrieved categories with pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          description: The unique identifier for the category.
                          example: 1
                        name:
                          type: string
                          description: The name of the category.
                          example: Technology
                        slug:
                          type: string
                          description: The unique slug for the category.
                          example: technology
                        description:
                          type: string
                          description: The description of the category (optional).
                          example: Articles related to technology and innovation.
                          nullable: true
                        createdAt:
                          type: string
                          format: date-time
                          description: The date and time when the category was created.
                          example: 2025-06-05T07:55:00Z
                        updatedAt:
                          type: string
                          format: date-time
                          description: The date and time when the category was last updated.
                          example: 2025-06-05T07:55:00Z
                        deletedAt:
                          type: string
                          format: date-time
                          description: The date and time when the category was deleted (optional).
                          example: null
                          nullable: true
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                        example: 1
                      limit:
                        type: integer
                        example: 10
                      showPerPage:
                        type: integer
                        example: 10
                      totalCategories:
                        type: integer
                        example: 50
                      totalPages:
                        type: integer
                        example: 5
        '400':
          description: Validation errors or malformed request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnAuthorizedError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
...
  /api/admin/categories/delete-category:
    delete:
      summary: Delete categories by IDs
      description: Deletes categories by their IDs (soft delete, sets deletedAt).
      operationId: adminDeleteCategories
      tags:
        - Admin
      security:
        - jwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - ids
              properties:
                ids:
                  type: array
                  items:
                    type: integer
                  example:
                    - 1
                    - 2
      responses:
        '200':
          description: Categories deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          description: Validation errors or malformed request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnAuthorizedError'
        '404':
          description: Some categories not found with the provided IDs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
...
  /api/admin/tags/get-tags:
    post:
      summary: Fetch tags with pagination and search
      description: Retrieves tags from the database with pagination and optional search.
      operationId: adminGetTags
      tags:
        - Admin
      security:
        - jwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                page:
                  type: integer
                  example: 1
                limit:
                  type: integer
                  example: 10
                search:
                  type: string
                  example: tech
      responses:
        '200':
          description: Successfully retrieved tags with pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          description: The unique identifier for the tag.
                          example: 1
                        name:
                          type: string
                          description: The name of the tag.
                          example: Tech
                        createdAt:
                          type: string
                          format: date-time
                          description: The date and time when the tag was created.
                          example: 2025-06-05T07:55:00Z
                        updatedAt:
                          type: string
                          format: date-time
                          description: The date and time when the tag was last updated.
                          example: 2025-06-05T07:55:00Z
                        deletedAt:
                          type: string
                          format: date-time
                          description: The date and time when the tag was deleted (optional).
                          example: null
                          nullable: true
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                        example: 1
                      limit:
                        type: integer
                        example: 10
                      showPerPage:
                        type: integer
                        example: 10
                      totalTags:
                        type: integer
                        example: 50
                      totalPages:
                        type: integer
                        example: 5
        '400':
          description: Validation errors or malformed request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnAuthorizedError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
...